Update isPseudoRealReg and isRealReg in OMRRealRegister
These two queries are stale and currently consider vector registers
to be PseudoRealRegisters, when they should be RealRegisters. This
commit fixes that issue.

The commit also adds a helpful comment inside RealRegisterEnum.hpp to
indicate that the VRF and FPR bases are the same because VRF0-15 overlap
with FPR0-15 and so are initialized the same.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>