Add relocation records correctly inside ConstantDataSnippets
`TR_SpecialRamMethodConst` and similar types do not need to be relocated
for remote compilations. This commit avoids adding relocations for those
data types.

`TR_JNISpecialTargetAddress` and similar types see crashes or AOT relocation
failures. This commit adds a proper query so that the relocation record
is only generated when needed and is correct to do so.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>