Handle specific arraycopy case
The arraycopy codegen was not expecting a specific pattern
where the second aladd had a NULL address child and an lsub
integer child. Since the address child is NULL we use the
integer child as the base register, but the code for cleaning
up the folded lsub assumed the index register was used. A simple
fix which selects the appropriate register in the memref
handles this case.

Signed-off-by: Simon Hirst <shirst@ca.ibm.com>