Add JITServer checks
We add JITServer checks to prevent the creation of TOC entries
and local relocations. Note: the local relocation is problematic because
when we do a relocation on the client, if there is already data in the relo location
then we end up with garbage. This is becase we expect the relo location to be populated with
zeroes and OR the new data in. Not sure if this is the best solution.

Note: the TOC entry change might be unnecessary (see comment near it)
Note2: OMR doesn't know about JITServer.. so we probably can't add JITServer checks here..
Will need something else.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>