Don't emit data for remote compilations
For remote compilations, we must not emit the actual
address when generating a fixed sequence on power during
a remote compilation. This is because during relocation time
on the client, the new data will be OR'ed in. This infrastructure
depends on the fact that the data in the patch location is
initially zero.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>