Correctly add relocations for addressConstants
This commit stops using the needsRelocationsForStatics query
for startPC and J9Method data types because it doesn't return the
correct result on Power. On X, perhaps we don't need to relocate
those data types, but on Power we do. So we simplify the check and
broaden its range. We can come back to this and decide what other query
to use instead.

We also update the ConstantDataSnippet's emitAddressConstant routine.
This thing needs a relocation (and is used by the startPC check above).

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>