This commit is not ready for prime time yet
But it's documenting an assert I used to detect if
we ever generate TOC code for a remote compilation.

This thing fires, but we don't see any errors in the wild
related to it without this assert.. and getting a log
was proving to be impossible. So I will come back and investigate
this later.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>