Generate relocatable code when using the loadAddressConstant query for JITServer compilations.
We would generate the lookup() IL which will map to scheme 3 and 4.
These schemes use loadAddressConstant.

For cold runs, we wouldn't see this issue. I'm guessing because we would
do an AOT compile, and so we would end up generating relocatable code because
loadAddressConstant will generate relo code if comp->reloCode query is on.

But during the warm run, the AOT load would fail. At this point
we would do a regular jit compile remotely. This would be a problem
because the lookup table evaluation would now not generate relo code.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>