Define subclass TR::map of std::map to make maps easier to declare
​
STL map types have been extremely verbose because of the typed_allocator
of pairs and because the allocator can't be specified without also
explicitly specifying the ordering comparison. This verbosity has led to
the widespread use of typedefs for multiple related types when trying to
name an instantiation of std::map. Here's an example:
​
    typedef TR::typed_allocator<std::pair<uint32_t const, uint32_t>, TR::Region &> InfoMapAllocator;
    typedef std::less<uint32_t> InfoMapComparator;
    typedef std::map<uint32_t, uint32_t, InfoMapComparator, InfoMapAllocator> InfoMap;
    InfoMap _blockMap;
​
Writing all of this every time has also been a source of bugs. In
particular, the const qualifier on the first type in the pair is
required, but it's easy to accidentally leave it out and have things
appear to work.
​
The allocator defaults to TR::Region&. This differs from the default for
TR::vector, but it's what most maps use.
​
With the TR::map class defined in this commit, the above could be
written instead as:
​
    TR::map<uint32_t, uint32_t> _blockMap;